ACL assertion should never exist on 'read' privilege
This is because all read assertion must be implemented as ACL filter
to be sure that things are correctly filtered when using getters.

Duplicate the logic with ACL assertions is counter-productive, because
it should have already been filtered before reaching that point. And if
it wasn't done correctly, then the ACL filter must be changed, never
the assertation.