Prevent Chrome to reload payment iframe #6272
Chrome, but not Firefox, would show a warning to the user to re-submit
data when use goes back in history  because the iframe is loaded via POST.
This does not make any sense for the user as he never submitted anything himself.

So to prevent that, we manipulate the browser history. After the iframe is loaded
via POST we inject a fake history state. This allow to go back normally to that
fake state and intercept the case to go back one state further, effectively
skipping over the POST request.