Replace lint-staged with bash
lint-staged provided nothing than we cannot do in bash, and it hid errors
that could happen in `./ng-lint-staged.js`, making it impossible to commit properly.

It also linted twice. Once with the main project and once with the e2e. But
it did so each time while overwriting the tsconfig, so effectively doing
the exact same work twice for benefit at all and reported inconsistent things.

Timing are:

- `time yarn lint` ~53s
- `time yarn lint okpilot` ~53s (same !)
- `time ./node_modules/.bin/lint-staged` ~24s
- `time ./bin/pre-commit.sh` ~13s

Important to note is that we do not lint e2e code, but we never did. Once we start
actually writing e2e we could add a section to lint the e2e project.

Also we do not lint HTML, and we never did, but we probably should.