Cached account.total_balance for group account #7149
This allow for sorting and easy filtering on the balance of group
account. Previously those the total balance for group account were
computed live and thus were not available for sorting and filtering.

We now store them in DB and compute them via triggers.

This also assert that a group account can never have a non-zero balance.

To make into production, we will have to use:

```sh
./vendor/bin/doctrine-dbal dbal:run-sql 'CALL update_account_balance(10000);'
```