Atomic builds
It's important to always have htdocs/index.php available, so that already
loaded clients can still communicate with GraphQL API even during builds.

So we copy that file at the very beginning of any kind of builds to be
available as soon as possible. And we remove it only a very very short moment
when the build is complete and ready to be moved to final destination.

And also the client should be available to be loaded even if a new client is
currently building. So the whole build process is done in a separate folder
and only when everything is ready we `mv` it instantly to its final destination.