SelectAdminApprovedModal: show when a course/welcomeSession is fully booked #7313
We actually show when the number of bookings on the admin_approved bookable
reached the simultaneous bookings max, not when the actual admin_only bookable
used for confirmed subscriptions reached its limit. This may not be always
accurate if the admin forgot to delete denied bookings or registered
someone manually to the course.