getSharedBookings have no more restrictions on bookable tags and type #7313
In a first step, this method was first used for storage usage only. Restrictions where there to prevent to do to much queries.
A second step enlarged usage to formations and welcome sessions (more tags)
Actually, the third step is to use similar feature for admin-aproved bookables too (more type)

In fact, this feature could be interesting for every bookable and can be enlarged to all tags and types.
The UsageServiceBookable is the determinant element to use that getter. Should not be used on regular listings.