Use int instead of Money for BookingCount #7445
When creating a custom operator we must be super-careful so that its name
does not duplicate one of the standar operators.

In this case when changing the operator type from Money to int, the final
operator name in GraphQL (visible in GraphiQl) was `EqualOperatorInt`. This
happens to be the same name as the standard, built-in, operator for existing
fields in DB of type int.

So when executing the query `graphql-doctrine` incorrectly use the standard
operator and thus attempted to read a field in DB that does not exists.

The solution is to change make sure that the final operator name in GraphQL is
unique. As the name is build from the class names, we must pseudo-namespace the
class name.