User cannot own himself #7928
A user does not need to, and cannot anymore, own himself
to declare that his a family owner.

This simplify our logic to filter family owner. That's all
users without owner. Thus this simple logic could be used
in the client without a risk to forget to also check `owner = myself`.

Also unit test for `setOwner` that are not specific to a User
are moved to their dedicated class to avoid confusion, and better
focus on either the generic `setOwner` or the specific `setOwner`
for `User`.