Untangle `RegisterComponent` and `RegisterConfirmComponent` #8117
`step` cannot be arbitrarily set from routing, because each specific `step` logic
is split in each specific component. So one given component is **always** step 1
(`RegisterComponent`), and the other component is **always** step 2 (`RegisterConfirmComponent`).

After registration is sent the email field is also hidden (as was already done for other fields).
Because there is no way to do anything else on the page anymore. Leaving the (editable) field made it
look like a half-broken page.

RegisterComponent is not a standard detail page. It fetches no data, it resolves no data,
and it has only one mutation which is custom and "hardcoded". Ans so it does not extend
`NaturalAbstractDetail` anymore.

Almost same goes for RegisterConfirmComponent. It does use a resolver, but it is a custom resolver
specifically for this component and does not use standard workflow of detail page. The only thing
copied from `NaturalAbstractDetail` is `this.route.data.subscribe` for the resolver, which is minimal.

`AnonymousService` and `NewService` are dropped altogether because they both had only a single
purpose, to build a form with validators. This can be written in a much more straightforward
way with `FormBuilder`. And this keep our custom needs centralized where they are needed with the
code flowing in a single class, instead of three.

This also avoid creating the same form for both components and trying to adapt it for each components.
Instead each components build directly exactly what it needs, nothing more, nothing less.

And finally we better flow our API types from UserByTokenResolver to the form itself.