Prevent layout thrashing in sidebars #7430
As documented in https://material.angular.io/components/sidenav/api#MatDrawerContainer,
the `autosize` feature is dangerous and costly. It actually cause layout thrashing
in Natural demo, as well as Theodia, and possibly other projects. Removing it show
no regression at all, and the menu width still adapt enough to its content.

If it happens in the future that the menu width needs to be recomputed, the `autosize`
must **not** be re-introduced, but instead we should expose
`MatDrawerContainer.updateContentMargins()` on our `NaturalSidenavContainerComponent`,
and thus allowing re-computation on specific events, and not on every single Angular digest.