Avoid possible mis-usages and tighten types #8155
This remove all existing `any` in the component and instead rely
on natural type flow to get strongly typed component.

While doing so, we fix a bug where a booking without owner would have
crashed the app.

This also reduce mis-usage by passing two incoherent `query` and `mapResultFunction`. There is only one input, and the details are internal
to the component.

It also correct typing from API which were incorrectly declared but was
not erroring because of usage of `any`.