Do not force an user without a password through the registration process #8283
Maybe the user was added manually by an administrator or an existing member to
his household and should not have to register to recover access to his account.
A new member is forced to confirm his account (which creates mandatory bookings
and account) and choose a login before being able to ask a password reset.