Ignore already imported transaction line #8306
Turns out a CAMT can contains the exact same transaction line
multiple times over time. Until now we rejected those duplicates,
assuming that each new CAMT file would never have duplicates,
and that finding one would mean that the exact same file was
being imported twice (by human error).

But our assumption was incorrect. Over 4'491 lines imported
across all shops, over the last two years, we identified 170
exact and legitimate duplicates. From now on we silently ignore
a line that was already imported.

In addition to that, "duplicate" was incorrectly defined. We assumed
that our `importedId`, more specifically the `<EndToEndID>`, was
universally unique. But it is not the case (!). It is possible to see
the same `<EndToEndID>` at different dates, even months appart. I assume
this means that an automatic monthly payment was setup and the bank
consider each monthly payment as part of the same `<EndToEndID>`.

For us that meant that monthly payment (for one specific user) were always
rejected and had to be credited manually. From now on, we include the
transaction date in the existence check, so that is is possible to import
recurring payments.