Support Raiffeisen CAMT files #8832
Raiffeisen CAMT files may have a single `<Ntry>` that contains multiple
`<TxDtls>` with the same duplicated `<AcctSvcrRef>`. I *think* this is
against the CAMT spec, but we must support it anyway.

We try hard to keep retro-compatibility with pre-existing `importedId`,
so we detect whether we have a Raiffeisen case. And if we do, then we
append an arbitrary suffix to the `<AcctSvcrRef>`.

The suffix is "-epicerio-$index" where `$index` is the index in the
array of `<TxDtls>`. That means that if for some reason Raiffeisen
generates a new CAMT with the same `<Ntry>` but a different order of
`<TxDtls>`, then Epicerio will incorrectly re-import again the same
transaction. That means Epicerio would credit to members twice the
amount. That is obviously very bad, but we can only hope that Raiffeisen
would not do that.