Don't show incorrect items after navigateTo #9054
If the navigateTo navigates to the same page but different ID, eg: from
/folder/1 to /folder/2, then Angular will recycle the exact same
instance of selector. And the click event used to navigate will keep
propagating until it gives focus to the "new" selector which will
trigger opening the dialog to select an item.

First we should not open a dialog when we only want to navigate. But
most importantly the "new" dialog will actually open with "old" options
leading to potentially incorrect items to be shown and selected. This
would typically end up in the server rejecting the mutation.