Update date is equal to creation date #9021
When creating something the update date is also set, instead of being left to
`null`. This matches what is typically done on filesystem. And most importantly
it allow for huge optimization when sorting things by their latest update, since
we do not need to compute "either creation date or update date". That allows
MariaDB DB to optimize the query, and even more so if we add an index on that
single column.