When creating transaction preserve lines even if server errors #9297
When trying to create an invalid transaction the server will reject it
with an error. In that case the transaction lines used to disappear from
the GUI, so end-user could not easily fix it to re-submit. Instead, he
had to re-enter everything manually.

Now we keep the line as editable until we are sure that the mutation
was successful. So we can re-submit them if need be.