Bookable can have a waiting list #9639
If a bookable has no waiting list, the number of booking is strictly
limited by `simultaneousBookingMaximum`. Regardless if the booking is
`application`, `booked` or `processed`.

If a bookable has a non-zero `waitingListLength`, then it is possible to
create booking of status `application` (and only that one) over the
limit of `simultaneousBookingMaximum`, and up until
`simultaneousBookingMaximum + waitingListLength`.

In the GUI we previously showed the count of only non-application
bookings. Now we show the count for all bookings. So end-user can
actually understand whether a bookable is fully booked or not.

We also show both limits separately, so end-user has a chance to
understand the system, especially with the new tooltip explaining things.
It is important to note that in `simultaneousBookingMaximum` limit (the
first number), it can be a mixed of application or non-application
bookings. There is no way to know count of applications and the count of
non-applications without going in admin.