Show either availability or usage status #9639
By having two different columns for each, we can better cater to the
need of each use-case, even showing both columns at the same time for
curious people.

Also the new `avoidTooManySqlQueries` is reverted to be based on the
`Bookable` type because NFC courses have no booking limits, but we still
want to show availability and usage status for them.