Filter users by absence of booking's bookable's tags #9707
It was possible to say "users with a booking whose bookable's does not
have this tag". It worked well if the bookable had exactly 1 tag. But
if the bookable had more tags, then it would incorrectly be selected
anyway, because the other tags would fulfill the "not have this tag"
part.

Now, if the bookable tags include at least "this tag" (and others), then
the whole bookable is excluded.