Can combine users query and transactionLines query in single XHR #9781
When creating a custom operator, we must be super-careful so that its name
does not duplicate one of the standard operators.

In this case, the final operator name in GraphQL (visible in GraphiQl) was
`EqualOperatorMoney`. This happens to be the same name as the standard,
built-in, operator for existing fields in DB of type Money.

So when executing the query `graphql-doctrine` incorrectly used the custom
operator and thus attempted to read a field in DB that does not exists.

The solution is to change make sure that the final operator name in GraphQL is
unique. As the name is build from the class names, we must pseudo-namespace the
class name.