HasBooking* operators share the same booking #9808
Most HasBooking* operators filter on the same booking. So we can
cumulate filter on the same booking to say "all users who have at least
one booking that is terminated and that is a SUP".

But HasBookingWithTaggedBookable has a variant with the new `sameBooking`
variable that can be set to false in order to say "all users who have at
least one booking that is terminated and that also have any other
booking is a SUP".