Support `?` and unicode in avatar #10077
On a total of 13'150 names in production data, the new initial algorithm
yields 1'078 (8%) different results from before 2d849505c771. Of those
differences, most are only a change in length, because it was unlimited
but now defaults to 2 characters.

The rest are 53 (0.4%) differences that are about special characters. A
manual review of those 53 cases shows that they are mostly clear
improvement, or at least not a downgrade, eg:

For `Alice / Bob / Charlie`:

```diff
- A/B/C
+ AB
```

For `John (and Jane) Doe`:

```diff
- J(JD
+ JD
```