Update eslint rules #10134
It's been a while since we updated our rules. This time we opt in into
all rules and explicitly disable the ones that don't fit our code. This
way we will automatically receive new rules in the future, and we can
decide whether we accept or reject them, instead of needing to re-check
which one is new at every update.