Don't over compute account balance #11178
We used to always re-computing both old/new, and both
debit/credit accounts for any changes. So that whenever we update a
transaction lines, we ran update_account_balance exactly 4 times, even
if we only marked the line as reconciled.

With this change, we are much more clever about it:

- Only line balance or account changes may trigger computation, all
other changes will do nothing
- Swapping both accounts with each other will compute each account 1
time, instead of 2 times

So we run account balance computation 0, 1, or 2 times per transaction
lines.