Don't reuse stale cache for ACL filer #11390
FilterCollection must be aware that our ACL filter internal state has
change whenever we change user. So that the Doctrine is able to properly
handle its own cache and not re-use an incorrect cache entry for the
wrong user.

This should not be a huge deal in real, production, use-cases. But it
very easily breaks unit tests where we change user often within the same
process.